Feat/campus timetable 4

+++ server/src/models/campus/timetablesets.model.ts
@@ -14,28 +14,30 @@ export interface TimetableSetListResponse {
 }
 
 export const TimetableSetsModel = {
-  getAllSets: async (): Promise<any[]> => {
+  getAllSets: async (userId : number): Promise<any[]> => {
     const [rows] = await pool.query(
       `SELECT 
-        timetable_list_id AS timetable_list_id,
-        timetable_name AS timetable_name
+        timetable_list_id,
+        timetable_name
       FROM timetable_list
-      ORDER BY timetable_list_id`
+      WHERE user_id = ?
+      ORDER BY timetable_list_id`,
+      [userId]
     );
     return rows as any[];
   },
 
-  createSet: async (name: string): Promise<void> => {
+  createSet: async (name: string, userId : number): Promise<void> => {
     await pool.query(
-      "INSERT INTO timetable_list (timetable_name, user_id) VALUES (?, 2)",
-      [name]
+      "INSERT INTO timetable_list (timetable_name, user_id) VALUES (?,?)",
+      [name, userId]
     );
   },
 
-  deleteSet: async (id: number): Promise<void> => {
+  deleteSet: async (id: number, userId : number): Promise<void> => {
     await pool.query(
-      "DELETE FROM timetable_list WHERE timetable_list_id = ?",
-      [id]
+      "DELETE FROM timetable_list WHERE timetable_list_id = ? AND user_id = ?",
+      [id, userId]
     );
   },
 };
